+++ x/evm/statedb/journal_test.go
@@ -2,6 +2,7 @@ package statedb_test
 
 import (
 	"fmt"
+	"math/big"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/core/vm"
@@ -28,11 +29,7 @@ func (s *Suite) TestPrecompileSnapshots() {
 	s.T().Log("Set up helloworldcounter.wasm")
 
 	wasmContract := precompiletest.SetupWasmContracts(&deps, &s.Suite)[1]
-	type Transition struct {
-		Run                 func(deps *evmtest.TestDeps) *vm.EVM
-		AssertionsBeforeRun func(deps *evmtest.TestDeps)
-	}
-	fmt.Printf("wasmContract: %v\n", wasmContract)
+	fmt.Printf("wasmContract: %s\n", wasmContract)
 	assertionsBeforeRun := func(deps *evmtest.TestDeps) {
 		precompiletest.AssertWasmCounterState(
 			&s.Suite, *deps, wasmContract, 0,
@@ -63,30 +60,34 @@ func (s *Suite) TestPrecompileSnapshots() {
 	)
 	s.Require().NoError(err, deployResp)
 
-	deps.EvmKeeper.ERC20().Mint()
-	_, _, err := deps.EvmKeeper.ERC20().Mint(
+	// deps.EvmKeeper.ERC20().Mint()
+	contract := deployResp.ContractAddr
+	_, evmObj, err := deps.EvmKeeper.ERC20().Mint(
 		contract, deps.Sender.EthAddr, evm.EVM_MODULE_ADDRESS,
 		big.NewInt(69_420), deps.Ctx,
 	)
+	s.Require().NoError(err)
 
-	evmtest.TransferWei()
+	stateDB := evmObj.StateDB.(*statedb.StateDB)
+	s.Equal(50, stateDB.Journal.DirtiesLen())
+	// evmtest.TransferWei()
 
 	s.T().Log("Run state transition")
 
-	evmObj := run(&deps)
+	evmObj = run(&deps)
 	stateDB, ok := evmObj.StateDB.(*statedb.StateDB)
 	s.Require().True(ok, "error retrieving StateDB from the EVM")
 	s.Equal(0, stateDB.Journal.DirtiesLen())
 
-	ctxBefore, _ := deps.Ctx.CacheContext()
+	_, _ = deps.Ctx.CacheContext()
 	assertionsAfterRun(&deps)
-	err := stateDB.Commit()
+	err = stateDB.Commit()
 	s.NoError(err)
 	assertionsAfterRun(&deps)
 
 	s.Equal(0, stateDB.Journal.DirtiesLen())
 
-	s.Require().EqualValues(ctxBefore, deps.Ctx,
-		"StateDB should have been committed by the precompile",
-	)
+	// s.Require().EqualValues(ctxBefore, deps.Ctx,
+	// 	"StateDB should have been committed by the precompile",
+	// )
 }